+++ libjggtoolsbase/code/src/integer.cpp
@@ -20,7 +20,7 @@ namespace NSCLP {
    }
    string Integer::toString() { return to_string(numberValue); }
    char*  Integer::toChar  (char *buff, size_t size) {
-        int res = snprintf(buff, size, "%d", (int) numberValue);
+        size_t res = snprintf(buff, size, "%d", (int) numberValue);
         if (res < 0 || res >= size ) throw new ToolsOutOfSpaceException(size);
         return buff;
    }
